Revert "Resolve dependency loop"
This reverts commit 2badae38a626dc665dd91a2ea6f7041b1661692d.
