+++ mcmd/version/molgenis_version.py
@@ -5,8 +5,7 @@ import requests
 from requests import HTTPError
 
 from mcmd.client import api
-from mcmd.client.errors import MolgenisOfflineError
-from mcmd.utils.errors import McmdError
+from mcmd.utils.errors import McmdError, MolgenisOfflineError
 
 _version_number = None
 _version = None
